requirements.txt
chardet==3.0.4
click==7.1.2
cmake==3.18.2.post1
decorator==4.4.2
face-recognition==1.3.0
face-recognition-models==0.3.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
moviepy==1.0.3
numpy
pandas
opencv-python
Pillow==8.0.1
proglog==0.1.9
requests==2.24.0
tqdm==4.51.0
urllib3==1.25.11
wincertstore==0.2
dlib
datetime
plotly

